Improve identifier generalization and camel casing.
Rather than blindly capitalizing the first letter of each word (and
leaving the rest of the word lowercase), we now only perform
capitalization if the word is already fully lowercase; otherwise, we
assume that the casing is intentional.

Additionally, the first word of generated methods identifiers are no
longer lowercased.